+++ clarity/src/vm/tests/datamaps.rs
@@ -661,7 +661,10 @@ fn bad_define_maps() {
         RuntimeCheckErrorKind::ExpectsAcceptable("Unknown type name: contents".to_string()).into(),
         RuntimeCheckErrorKind::ExpectsAcceptable("Expected name".to_string()).into(),
         RuntimeCheckErrorKind::IncorrectArgumentCount(3, 4).into(),
-        RuntimeCheckErrorKind::ExpectsAcceptable("Invalid type description".to_string()).into(),
+        RuntimeCheckErrorKind::ExpectsAcceptable(
+            "Unexpected error type during runtime analysis: InvalidTypeDescription".to_string(),
+        )
+        .into(),
     ];
 
     for (test, expected_err) in tests.iter().zip(expected.into_iter()) {