+++ stacks-common/src/types/mod.rs
@@ -618,6 +618,22 @@ impl StacksEpochId {
         self < &StacksEpochId::Epoch25
     }
 
+    /// Whether `from-consensus-buff` treats unexpected serialization as `none` or causes
+    /// an error that makes the transaction un-includable in a block.
+    pub fn treats_unexpected_serialization_as_none(&self) -> bool {
+        self >= &StacksEpochId::Epoch34
+    }
+
+    /// Whether or not this epoch rejects `SupertypeTooLarge` errors.
+    pub fn rejects_supertype_too_large(&self) -> bool {
+        self < &StacksEpochId::Epoch34
+    }
+
+    /// Whether or not this epoch rejects parse-depth errors.
+    pub fn rejects_parse_depth_errors(&self) -> bool {
+        self < &StacksEpochId::Epoch34
+    }
+
     /// What is the sortition mining commitment window for this epoch?
     pub fn mining_commitment_window(&self) -> u8 {
         MINING_COMMITMENT_WINDOW
@@ -877,6 +893,24 @@ impl StacksEpochId {
             | StacksEpochId::Epoch31
             | StacksEpochId::Epoch32 => false,
             StacksEpochId::Epoch33 | StacksEpochId::Epoch34 => true,
+        }
+    }
+
+    pub fn handles_with_stx_combined_check(&self) -> bool {
+        match self {
+            StacksEpochId::Epoch10
+            | StacksEpochId::Epoch20
+            | StacksEpochId::Epoch2_05
+            | StacksEpochId::Epoch21
+            | StacksEpochId::Epoch22
+            | StacksEpochId::Epoch23
+            | StacksEpochId::Epoch24
+            | StacksEpochId::Epoch25
+            | StacksEpochId::Epoch30
+            | StacksEpochId::Epoch31
+            | StacksEpochId::Epoch32
+            | StacksEpochId::Epoch33 => false,
+            StacksEpochId::Epoch34 => true,
         }
     }